Refactor
This is a complete overhaul of the previous code. While the previous
code got the job done, it was a bit of a headache to maintain and seemed
prohibitive to community involvement. With this refactor, the code has
been subdivided into modules with reduced responsibility. JSHint was
added to give some semblance of uniformity to how the code is written.

In addition to refactoring, the code now contains a full suite of unit
tests to reinforce that future code changes don't unexpectedly break any
existing functionality. This should aid in allowing the community to
confidently make changes.

Build monitoring has also been added via Travis-CI to ensure that the
code works on a myriad of Node.js and IO.js versions.

Build monitoring is augmented further using Istanbul with Coveralls to
report code coverage statistics.

Build Report: https://travis-ci.org/dgarlitt/karma-nyan-reporter
Coverage report: https://coveralls.io/r/dgarlitt/karma-nyan-reporter